server/requirements.txt
fastapi
uvicorn
fastapi-mail
slowapi
pydantic
python-frontmatter>=1.0.0